refactor: organize files into layers folders
