+++ src/get-saml-metadata/src/data/adapters/FileValidatorAdapter.ts
@@ -1,5 +1,5 @@
 import { existsSync } from 'fs'
-import { BaseFileValidator } from '../../domain/protocols/BaseFileValidator'
+import { BaseFileValidator } from '../../use-cases/ports/BaseFileValidator'
 
 export class FileValidatorAdapter extends BaseFileValidator {
   isValid (filepath: string): boolean {
