--- src/get-saml-metadata/src/delivery/protocols/IController.ts
@@ -1,6 +1,0 @@
-import { IRequest } from './IRequest'
-import { IResponse } from './IResponse'
-
-export interface IController {
-  handle: (request: IRequest) => Promise<IResponse>
-}
