--- src/get-saml-metadata/src/delivery/protocols/IRequest.ts
@@ -1,3 +1,0 @@
-export interface IRequest {
-  request: any
-}
