--- src/get-saml-metadata/src/delivery/protocols/IResponse.ts
@@ -1,3 +1,0 @@
-export interface IResponse {
-  response: any
-}
