+++ src/get-saml-metadata/src/entities/IExternalData.ts
