+++ src/get-saml-metadata/src/entities/IMetadataTypes.ts
