+++ src/get-saml-metadata/src/entities/errors/InvalidPathOrUrlError.ts
