+++ src/get-saml-metadata/src/entities/types/Entity.ts
