+++ src/get-saml-metadata/src/entities/types/ValueObject.ts
