+++ src/get-saml-metadata/src/interface-adapters/factories/makeFileLoaderAdapter.ts
