+++ src/get-saml-metadata/src/interface-adapters/utils/ExternalDataMapper.spec.ts
