+++ src/get-saml-metadata/src/use-cases/GetExternalDataInteractor.spec.ts
@@ -1,17 +1,17 @@
-import { fakeMetadata, validMetadataString, validXmlMetadata } from '../../../../testdata/fakes'
-import { makeXmlMetadata } from '../factories/makeXmlMetadata'
-import { IXmlMetadataLoaderRepository } from '../utils/IXmlMetadataLoaderRepository'
-import { XmlMetadata } from '../value-objects/XmlMetadata'
-import { IExternalDataMapper } from '../../application/protocols/IExternalDataMapper'
-import { IMetadata } from '../../application/protocols/IMetadataTypes'
+import { fakeMetadata, validMetadataString, validXmlMetadata } from '../../../testdata/fakes'
+import { makeXmlMetadata } from '../data/factories/makeXmlMetadata'
+import { IExternalData } from '../entities/IExternalData'
+import { IMetadata } from '../entities/IMetadataTypes'
+import { XmlMetadata } from '../entities/value-objects/XmlMetadata'
 import { GetExternalDataInteractor } from './GetExternalDataInteractor'
 import { GetExternalDataRequestModel } from './GetExternalDataRequestModel'
 import { GetExternalDataResponseModel } from './GetExternalDataResponseModel'
 import { IGetExternalDataInputBoundary } from './IGetExternalDataInputBoundary'
 import { IGetExternalDataOutputBoundary } from './IGetExternalDataOutputBoundary'
 import { IResponseModel } from './IResponseModel'
-import { IExternalData } from '../../application/protocols/IExternalData'
-import { IMetadataMapper } from '../../application/protocols/IMetadataMapper'
+import { IExternalDataMapper } from './ports/IExternalDataMapper'
+import { IMetadataMapper } from './ports/IMetadataMapper'
+import { IXmlMetadataLoaderRepository } from './ports/IXmlMetadataLoaderRepository'
 
 const makePresenter = (): IGetExternalDataOutputBoundary => {
   class PresenterStub implements IGetExternalDataOutputBoundary {
