+++ src/get-saml-metadata/src/use-cases/GetExternalDataResponseModel.ts
@@ -1,4 +1,4 @@
-import { IExternalData } from '../protocols/IExternalData'
+import { IExternalData } from '../domain/protocols/IExternalData'
 
 export type UrlOrPath = string
