+++ src/get-saml-metadata/src/use-cases/IGetExternalDataInputBoundary.ts
