+++ src/get-saml-metadata/src/use-cases/IGetExternalDataOutputBoundary.ts
