+++ src/get-saml-metadata/src/use-cases/ports/IExternalDataMapper.ts
