+++ src/get-saml-metadata/src/use-cases/ports/IValidator.ts
