+++ src/get-saml-metadata/src/use-cases/ports/IXmlMetadataLoaderRepository.ts
@@ -1,4 +1,4 @@
-import { XmlMetadata } from '../value-objects/XmlMetadata'
+import { XmlMetadata } from '../../entities/value-objects/XmlMetadata'
 
 export interface IXmlMetadataLoaderRepository {
   load: (urlOrPath: string) => XmlMetadata
